Create Random_Gene_Assortment.txt
I used this code in order to generate 2 sets of 154 random genes from our lab's single knockout collection of S. cerevisiae. This was done in order to compare the significance of hits I got from a genetic screen to two sets of 154 random genes to ensure that the hits I collected were statistically significant.